modify ping bug

+++ command_sender.h
@@ -92,8 +92,8 @@ private:
     long long illegal_packets_=0;
     long long timeout_packets_=0;
 
-    ssize_t varn_delay_ = 0;
-    ssize_t std_delay_ = 0;
+    uint64_t varn_delay_ = 0;
+    uint64_t std_delay_ = 0;
 };
 
 class SendCommandSender : public CommandSender
